ml_service/requirements.txt
flask
flask-cors
mediapipe==0.10.21
opencv-python
numpy==1.26.4
scikit-learn
Pillow